+++ packages/website/src/snippet/counterAutoCount.ts
@@ -36,7 +36,7 @@ const subscriptions = Subscription.makeSubscriptions(SubscriptionDeps)<
     modelToDependencies: model => ({
       isAutoCounting: model.isAutoCounting,
     }),
-    depsToStream: ({ isAutoCounting }) =>
+    dependenciesToStream: ({ isAutoCounting }) =>
       Stream.when(
         Stream.tick(Duration.seconds(1)).pipe(Stream.map(Ticked)),
         () => isAutoCounting,